successfully wrangled cycle tab into own class
